requirements.txt
python-telegram-bot
pydub
requests